GH-28 Renamed GridState to GridInfo
